+++ src/js/language/vi_VN.js
@@ -239,7 +239,8 @@
                 PK: 'Pakistan',
                 RO: 'Romania',
                 TH: 'Thái Lan',
-                US: 'Mỹ'
+                US: 'Mỹ',
+                VE: 'Venezuela'
             }
         },
         regexp: {